显示的结果为整数...

+++ script/plot/splashlogplot.py
@@ -289,7 +289,7 @@ if __name__ == "__main__" :
             #print yData
             #print "==========================================\n"
 
-            appPlotDataList.append(string.atof(yData[0]) / 1000)   #  每个进程的运行情况
+            appPlotDataList.append(int(string.atof(yData[0]) / 1000))   #  每个进程的运行情况
         print name, appPlotDataList
         #plotdata = SplashPlotData(name = app, xData = , yData = appPlotDataList, color = color)
         plotDataList.append(appPlotDataList)    # 每个name系统中各个进程app的运行情况, 每行对应一个系统, 每列对应一个进程
